Add functions to database updater
Previously the addToDatabase function (or whatever it was called)
could more unsolved to solved, increment a playcount, or add a new
puzzle to the appropriate table. Now it can also move from solved to
unsolved (if the program can no longer solve the puzzle) as well as
update the hardest move if an easier route was found. Typically,
this latter function is useful when adding new solvemethods that are
easier for humans than previously implemented techniques.

The database was then updated, moving methods solved with the
tempermental bowmanbingo function back to the unsolved table. Except,
presumably, the ones that could now be solved with recently added
techniques.