Fix Simple Coloring bug
Simple Coloring was meant to check all possible chain sets for
each value. Instead, it was building a chain from the first
random cell it checked and then moving on, when other chain sets
might have been possible. Now it checks for others. The unsolved
db was checked again and more were solved.

Other code cleanup was done as well.