Add text files of puzzles
Databases don't travel with machines, apparently. With that
in mind, it may be useful to write a script that creates
and seeds the database if it is not already present. It
shouldn't matter if the puzzles are being stored on a server
that is accessed by everyone, I guess. I don't know.

In any case, the target version has also been lowered, mostly
in a failed attempt to make this run on an earlier version of VS.

Finally, the gitignore file has been added and will hopefully
come with the rest of the project now.